refactor(react-wildcat-hot-reloader): decouple Socket listener from hot reloader
Make the hot reloader agnostic of sockets. Prep-work to allow custom event listener types (polling,
kafka, etc)
