+++ packages/react-wildcat/package.json
@@ -40,7 +40,7 @@
     "resolve-path": "^1.3.0",
     "slash": "^1.0.0",
     "spdy": "^3.0.0",
-    "ws": "^0.8.0"
+    "ws": "^0.8.1"
   },
   "scripts": {
     "test": "echo \"no test specified\" && exit 0"
